Use std::variant to implement DebugSink
